reworked for lab2 Added JWT auth and GitHub OAuth, protected routes and othed staff

+++ app/routers/user.py
@@ -2,9 +2,12 @@ from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 from app import crud, schemas
 from app.database import get_db
-
+from app.dependencies import get_current_user, admin_only
+from app.models import User as UserModel
+from app.dependencies import admin_only
 router = APIRouter(prefix="/users", tags=["users"])
 
+# Чтение открыто
 @router.get("/", response_model=list[schemas.UserOut])
 def read_users(db: Session = Depends(get_db)):
     return crud.get_users(db)
@@ -16,20 +19,44 @@ def read_user(user_id: int, db: Session = Depends(get_db)):
         raise HTTPException(status_code=404, detail="User not found")
     return user
 
+# Остальное protected
 @router.post("/", response_model=schemas.UserOut)
 def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
     return crud.create_user(db, user)
 
 @router.put("/{user_id}", response_model=schemas.UserOut)
-def update_user(user_id: int, user_update: schemas.UserUpdate, db: Session = Depends(get_db)):
+def update_user(user_id: int, user_update: schemas.UserUpdate, db: Session = Depends(get_db), current_user: UserModel = Depends(get_current_user)):
+    if current_user.id != user_id and not current_user.is_admin:
+        raise HTTPException(status_code=403, detail="Not authorized")
     user = crud.update_user(db, user_id, user_update)
     if not user:
         raise HTTPException(status_code=404, detail="User not found")
     return user
 
 @router.delete("/{user_id}")
-def delete_user(user_id: int, db: Session = Depends(get_db)):
+def delete_user(user_id: int, db: Session = Depends(get_db), current_user: UserModel = Depends(get_current_user)):
+    if current_user.id != user_id and not current_user.is_admin:
+        raise HTTPException(status_code=403, detail="Not authorized")
     user = crud.delete_user(db, user_id)
     if not user:
         raise HTTPException(status_code=404, detail="User not found")
     return {"detail": "User deleted"}
+
+# Админ может менять verified и admin флаги
+@router.patch("/{user_id}/role")
+def update_user_role(
+    user_id: int,
+    role_update: schemas.UserRoleUpdate,
+    db: Session = Depends(get_db),
+    current_user: UserModel = Depends(admin_only)
+):
+    user = crud.get_user(db, user_id)
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
+    if role_update.is_author_verified is not None:
+        user.is_author_verified = role_update.is_author_verified
+    if role_update.is_admin is not None:
+        user.is_admin = role_update.is_admin
+    db.commit()
+    db.refresh(user)
+    return {"detail": "User role updated"}
